Adding a script to rewrite the spritesheets.
It cannot read in the new sprite sheet definition yet, so it can only
convert old grid spritesheets.

There are many algorithms to put as many rectangles as possible in one
predefined rectangle. However we have another usecase:
We would like to get a fixed amount of rectangles into one rectangle, which
has a minimal size.
This can be solved using the algorithms to place as many rects into one
predefined rect by iterating over different sizes of the predefined rect
and trying to fit in all required rectangles.

The idea is based on the paper in
https://github.com/juj/RectangleBinPack/
The implementation is a python script, which calls the rectanglepacker as
found in my fork of that repository.